Allow sending of MMS as non-default app

+++ library/src/main/java/com/klinker/android/send_message/MmsSentReceiver.java
@@ -22,6 +22,7 @@ import android.content.Context;
 import android.content.Intent;
 import android.net.Uri;
 import android.provider.Telephony;
+import android.text.TextUtils;
 
 import com.google.android.mms.util_alt.SqliteWrapper;
 import com.klinker.android.logger.Log;
@@ -40,17 +41,22 @@ public class MmsSentReceiver extends BroadcastReceiver {
     public void onReceive(Context context, Intent intent) {
         Log.v(TAG, "MMS has finished sending, marking it as so in the database");
 
-        Uri uri = Uri.parse(intent.getStringExtra(EXTRA_CONTENT_URI));
-        Log.v(TAG, uri.toString());
+        String extraContentUri = intent.getStringExtra(EXTRA_CONTENT_URI);
+        if (!TextUtils.isEmpty(extraContentUri)) {
+            Uri uri = Uri.parse(extraContentUri);
+            Log.v(TAG, uri.toString());
 
-        ContentValues values = new ContentValues(1);
-        values.put(Telephony.Mms.MESSAGE_BOX, Telephony.Mms.MESSAGE_BOX_SENT);
-        SqliteWrapper.update(context, context.getContentResolver(), uri, values,
-                null, null);
+            ContentValues values = new ContentValues(1);
+            values.put(Telephony.Mms.MESSAGE_BOX, Telephony.Mms.MESSAGE_BOX_SENT);
+            SqliteWrapper.update(context, context.getContentResolver(), uri, values,
+                    null, null);
+        }
 
         String filePath = intent.getStringExtra(EXTRA_FILE_PATH);
-        Log.v(TAG, filePath);
-        new File(filePath).delete();
+        if (!TextUtils.isEmpty(filePath)) {
+            Log.v(TAG, filePath);
+            new File(filePath).delete();
+        }
     }
 
 }
